Overhauled AI System
The AI system has been completely replaced by a new system. The old
system used a lot of if statements to measure possible outcomes, the new
system is much more slick. It scores the possible moves, and the AI
reacts accordingly.
Next Step: Round out kinks in AI, make sure the logic is sound.
Replace the list with a queue